+++ twamp-client-c/src/report.c
@@ -351,8 +351,8 @@ int twamp_report(TWAMPReport *report, TWAMPParameters *param)
     if (!param->report_mode) {
         /* we need to serialize the root array, but we don't want to output its delimiters [ ],
          * and we need to omit the "," after the last member of the array */
-        int al = json_object_array_length(jo);
-        for (int i = 0; i < al ; i++) {
+        size_t al = json_object_array_length(jo);
+        for (size_t i = 0; i < al ; i++) {
             fprintf(stdout, "%s%s",
                       json_object_to_json_string_ext(json_object_array_get_idx(jo, i),
                                           JSON_C_TO_STRING_PRETTY | JSON_C_TO_STRING_SPACED),
